modify statistics tables

+++ jedi-consumer/src/main/java/xyz/hellothomas/jedi/consumer/domain/ExecutorTaskStatistics.java
@@ -42,12 +42,12 @@ public class ExecutorTaskStatistics {
     /**
      * 执行总数
      */
-    private Integer total;
+    private Long total;
 
     /**
      * 执行失败总数
      */
-    private Integer failure;
+    private Long failure;
 
     /**
      * 执行失败比例
@@ -57,22 +57,22 @@ public class ExecutorTaskStatistics {
     /**
      * 执行最长时间
      */
-    private Integer executionTimeMax;
+    private Long executionTimeMax;
 
     /**
      * 执行最短时间
      */
-    private Integer executionTimeMin;
+    private Long executionTimeMin;
 
     /**
      * 执行时间95线
      */
-    private Integer executionTimeLine95;
+    private Long executionTimeLine95;
 
     /**
      * 执行时间99线
      */
-    private Integer executionTimeLine99;
+    private Long executionTimeLine99;
 
     /**
      * 创建时间
